requirements.txt
langchain==0.1.0
langchain-experimental
sentence-transformers==2.3.0
qdrant-client==1.7.0
google-generativeai==0.3.0
beautifulsoup4==4.12.0
requests==2.31.0
selenium==4.16.0
schedule==1.2.0
apscheduler==3.10.0
python-telegram-bot==20.7
discord.py==2.3.0
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.0
